Extract agent action parser
